Fix #121: Make the dom package object a Scala object, not a js.Any.
This is a pretty major breaking change, as now `dom` does not
extend `Window` anymore. This means that selecting fields of the
window through `dom.someField` is broken. `dom.window.someField`
should be explicitly used instead.

We explicitly expose `document` as `dom.document`, because lots of
examples out there use it this way, and it seems reasonable to keep
that use case.