Add default values for RequestInit fields
This way, users do not have to recite every field when creating a RequestInit.
Also, the spec says the RequestInit dictionary must contain a field `window`
which is set to null. We might as well include this too as a final variable,
since the spec shows omitting it should be a type error.