Avoid forcing evaluation of crypto
Having `val crypto` in the package object means that when it's undefined
an exception is thrown immediately, which makes it impossible to use the
WebCrypto facade in such a case.

An example situation where this is preventative is using the WebCrypto
API from Node where its available as a library and not globally
available.