Move everything from `raw.*` to `dom.*`.
We introduce deprecated aliases from `raw.*` to the corresponding
elements of `dom.*`.

The short aliases in `html.*`, `svg.*`, `webgl.*` and `idb.*` are
kept.

The aliases in `webworkers.*` are deprecated like `raw.*` since
they are not shorter than the original names.

The aliases in `css.*` are deprecated because the object `css`
now conflicts with the DOM `CSS` global object on case-insensitive
file systems. The aliases are not really shorter than the original
names anyway.